Improved price consistency by Using Int for cents instead of Double and String

+++ server/Comanda/src/main/kotlin/dev/giuseppedarro/comanda/features/menu/data/MenuSchema.kt
@@ -14,7 +14,7 @@ object MenuItems : Table() {
     val id = varchar("id", 128)
     val categoryId = varchar("category_id", 128).references(MenuCategories.id)
     val name = varchar("name", 200)
-    val price = double("price")
+    val price = integer("price")
     val description = varchar("description", 500).nullable()
     val displayOrder = integer("display_order").default(0)
     val isAvailable = bool("is_available").default(true)
